requirements.txt
Flask_SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.1
Flask-Sockets
requests
flask_cors
Pillow >= 8.3.2